Support all Rails versions over Appraisal
