+++ .travis.yml
@@ -1,0 +1,12 @@
+gemfile:
+  - gemfiles/rails_3.0.gemfile
+  - gemfiles/rails_3.1.gemfile
+  - gemfiles/rails_3.2.gemfile
+  - gemfiles/rails_4.0.gemfile
+  - gemfiles/rails_4.1.gemfile
+  - gemfiles/rails_4.2.gemfile
+  - gemfiles/rails_5.0.gemfile
+  - gemfiles/rails_5.1.gemfile
+  - gemfiles/rails_5.2.gemfile
+  - gemfiles/rails_6.0.gemfile
+script: "bundle exec rake test"
