+++ Appraisals
@@ -1,0 +1,49 @@
+appraise "rails-3.0" do
+	gem "rails", "~> 3.0"
+	gem 'sqlite3', '~> 1.3', '< 1.4'
+end
+
+appraise "rails-3.1" do
+	gem "rails", "~> 3.1"
+	gem 'sqlite3', '~> 1.3', '< 1.4'
+end
+
+appraise "rails-3.2" do
+	gem "rails", "~> 3.2"
+	gem 'sqlite3', '~> 1.3', '< 1.4'
+end
+
+appraise "rails-4.0" do
+	gem "rails", "~> 4.0"
+	gem 'sqlite3', '~> 1.3', '< 1.4'
+end
+
+appraise "rails-4.1" do
+	gem "rails", "~> 4.1"
+	gem 'sqlite3', '~> 1.3', '< 1.4'
+end
+
+appraise "rails-4.2" do
+	gem "rails", "~> 4.2"
+	gem 'sqlite3', '~> 1.3', '< 1.4'
+end
+
+appraise "rails-5.0" do
+	gem "rails", "~> 5.0"
+	gem 'sqlite3', '~> 1.4'
+end
+
+appraise "rails-5.1" do
+	gem "rails", "~> 5.1"
+	gem 'sqlite3', '~> 1.4'
+end
+
+appraise "rails-5.2" do
+	gem "rails", "~> 5.2"
+	gem 'sqlite3', '~> 1.3.6'
+end
+
+appraise "rails-6.0" do
+	gem "rails", "~> 6.0"
+	gem 'sqlite3', '~> 1.4'
+end
