+++ gemfiles/rails_3.2.gemfile
@@ -1,0 +1,9 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "test-unit"
+gem "rails", "~> 3.2"
+gem "sqlite3", "~> 1.3", "< 1.4"
+
+gemspec path: "../"
