+++ gemfiles/rails_3.gemfile
@@ -1,0 +1,8 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "test-unit", "~> 3.0"
+gem "rails", "3.2.22.5"
+
+gemspec path: "../"
