+++ gemfiles/rails_4.gemfile
@@ -1,0 +1,8 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "test-unit", "~> 3.0"
+gem "rails", "4.2.11.1"
+
+gemspec path: "../"
