+++ gemfiles/rails_5.0.gemfile
@@ -1,0 +1,9 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "test-unit"
+gem "rails", "~> 5.0"
+gem "sqlite3", "~> 1.4"
+
+gemspec path: "../"
