+++ gemfiles/rails_5.2.gemfile
@@ -1,0 +1,9 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "test-unit"
+gem "rails", "~> 5.2"
+gem "sqlite3", "~> 1.3.6"
+
+gemspec path: "../"
