+++ gemfiles/rails_6.0.gemfile.lock
@@ -1,5 +1,5 @@
 PATH
-  remote: .
+  remote: ..
   specs:
     acts_as_slugable (1.0.2)
       rails
@@ -119,7 +119,7 @@ GEM
       method_source
       rake (>= 0.8.7)
       thor (>= 0.20.3, < 2.0)
-    rake (12.3.2)
+    rake (13.0.1)
     sprockets (4.0.0)
       concurrent-ruby (~> 1.0)
       rack (> 1, < 3)
@@ -145,8 +145,9 @@ PLATFORMS
 DEPENDENCIES
   acts_as_slugable!
   appraisal
+  rails (~> 6.0)
   rake
-  sqlite3
+  sqlite3 (~> 1.4)
   test-unit
 
 BUNDLED WITH
