+++ gemfiles/rails_6.gemfile
@@ -1,0 +1,8 @@
+# This file was generated by Appraisal
+
+source "https://rubygems.org"
+
+gem "test-unit", "~> 3.0"
+gem "rails", "6.0.1"
+
+gemspec path: "../"
