+++ gemfiles/rails_6.gemfile.lock
@@ -1,7 +1,7 @@
 PATH
-  remote: .
+  remote: ..
   specs:
-    acts_as_slugable (1.0.2)
+    acts_as_slugable (1.0.1)
       rails
 
 GEM
@@ -119,7 +119,7 @@ GEM
       method_source
       rake (>= 0.8.7)
       thor (>= 0.20.3, < 2.0)
-    rake (12.3.2)
+    rake (13.0.1)
     sprockets (4.0.0)
       concurrent-ruby (~> 1.0)
       rack (> 1, < 3)
@@ -127,7 +127,7 @@ GEM
       actionpack (>= 4.0)
       activesupport (>= 4.0)
       sprockets (>= 3.0.0)
-    sqlite3 (1.4.1)
+    sqlite3 (1.3.13)
     test-unit (3.2.9)
       power_assert
     thor (0.20.3)
@@ -145,9 +145,10 @@ PLATFORMS
 DEPENDENCIES
   acts_as_slugable!
   appraisal
+  rails (= 6.0.1)
   rake
   sqlite3
-  test-unit
+  test-unit (~> 3.0)
 
 BUNDLED WITH
    1.17.3
