+++ lib/acts_as_slugable/version.rb
@@ -1,3 +1,3 @@
 module Slugable
-  VERSION = "1.0.1"
+  VERSION = "1.0.2"
 end
