+++ test/fixtures/page.rb
@@ -1,4 +1,4 @@
 class Page < ActiveRecord::Base
-  validates :title, presence: true, length: { minimum: 2}
+  validates :title, presence: true, length: { minimum: 2 }
   acts_as_slugable source_column: :title, slug_column: :url_slug, scope: :parent
 end
